requirements.txt
anyio==4.2.0
APScheduler==3.10.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
load-dotenv==0.1.0
lxml==5.1.0
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2023.4
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
